Bring composer.lock inline with composer.json
You can't just do `composer update` here - that'll try and bring Symfony all the way up to 2.8, which this package isn't ready for!
Given that, I ran `composer update --prefer-lowest`.  The intention is to update the lock file further in future with newer versions, but this works for now.

I did temporarily lock doctrine/common on its original version to avoid that being downgraded for no reason.
I added the following to composer.json before running the update, but didn't need to leave it in, since this version-lock is already in the lock file:
"doctrine/common": "2.3.x-dev#bb0aebb"